+++ CHANGELOG.md
@@ -1,3 +1,40 @@
+<a name="3.6.2"></a>
+## [3.6.2](https://github.com/mengxiong10/vue2-datepicker/compare/v3.6.1...v3.6.2) (2020-08-10)
+
+
+### Bug Fixes
+
+* provide scoped slot to `input` slot ([#492](https://github.com/mengxiong10/vue2-datepicker/issues/492)) ([b3126a0](https://github.com/mengxiong10/vue2-datepicker/commit/b3126a0))
+* value entered manually in disabled range should be invalid ([#508](https://github.com/mengxiong10/vue2-datepicker/issues/508)) ([1bca35b](https://github.com/mengxiong10/vue2-datepicker/commit/1bca35b))
+
+
+
+<a name="3.6.1"></a>
+## [3.6.1](https://github.com/mengxiong10/vue2-datepicker/compare/v3.6.0...v3.6.1) (2020-07-13)
+
+
+### Bug Fixes
+
+* popup sholud be not auto closed in `time range` ([a4e5019](https://github.com/mengxiong10/vue2-datepicker/commit/a4e5019))
+
+
+
+<a name="3.6.0"></a>
+# [3.6.0](https://github.com/mengxiong10/vue2-datepicker/compare/v3.5.0...v3.6.0) (2020-06-07)
+
+
+### Bug Fixes
+
+* remove touchstart event ([#469](https://github.com/mengxiong10/vue2-datepicker/issues/469)) ([b6b770d](https://github.com/mengxiong10/vue2-datepicker/commit/b6b770d))
+
+
+### Features
+
+* add multi-select ([#444](https://github.com/mengxiong10/vue2-datepicker/issues/444)) ([7835ffb](https://github.com/mengxiong10/vue2-datepicker/commit/7835ffb))
+* add the prop `scroll-duration` ([#475](https://github.com/mengxiong10/vue2-datepicker/issues/475)) ([051ad89](https://github.com/mengxiong10/vue2-datepicker/commit/051ad89))
+
+
+
 <a name="3.5.0"></a>
 # [3.5.0](https://github.com/mengxiong10/vue2-datepicker/compare/v3.4.1...v3.5.0) (2020-04-22)